requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==24.2.0
certifi==2025.1.31
charset-normalizer==3.3.2
Django==3.2
django-htmx==1.19.0
django-model-utils==4.5.1
django-otp==1.3.0
django-render-partial==0.4
frozenlist==1.4.1
google-api-python-client==1.4.1
httplib2==0.22.0
idna==3.7
isodate==0.6.1
jalali_core==1.0.0
jdatetime==5.0.0
jsonfield==3.1.0
langdetect==1.0.9
lxml==5.3.0
material==0.1
msgpack==1.0.8
multidict==6.0.5
mysqlclient==2.2.4
oauth2client==4.1.3
packaging==24.1
pip-autoremove==0.10.0
pip-check-reqs==2.5.3
pipdeptree==2.23.1
platformdirs==4.2.2
purge==1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyotp==2.9.0
pyparsing==3.2.1
python-slugify==8.0.4
pytz==2024.1
redis==5.0.8
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rsa==4.9
six==1.16.0
sorl-thumbnail==12.10.0
sqlparse==0.4.4
swapper==1.4.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.2.2
vulture==2.11
yarl==1.9.4
zeep==4.2.1
